Connected edit user to backend
