+++ views/layout/partials/scripts.ejs
@@ -449,63 +449,4 @@
   }
 </script>
 
-<!-- Script for editing and deleting users -->
-<script>
-  async function submitUserEdits(id) {
-    var role, active;
-    if (
-      document.querySelector('input[name="role"]:checked').value &&
-      document.querySelector('input[name="active"]:checked').value
-    ) {
-      role = document.querySelector('input[name="role"]:checked').value;
-      active = document.querySelector('input[name="active"]:checked').value;
-    } else {
-      console.log('Something went wrong');
-    }
-    console.log(role, active);
-    const newPermissions = {
-      role: role,
-      enable: active,
-    };
 
-    try {
-      const editResponse = await fetch(`/users/editUser/${id}`, {
-        method: 'POST',
-        credentials: 'same-origin',
-        body: JSON.stringify(newPermissions),
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
-      if (editResponse.status === 200) {
-        console.log('wow it worked');
-      }
-    } catch (err) {
-      console.log('it didnt work:' + err);
-    }
-  }
-
-  function editUserPopup() {
-    $('#modalContent').hide();
-    $('#changeUserPermission').show();
-  }
-
-  function deleteUserPopup(idNum, email) {
-    if (confirm('Are You sure you want to delete this user?')) {
-      // delete the user of idNum
-      return new Promise((resolve, reject) => {
-        axios
-          .delete('/users')
-          .send({ userId: idNum, email: email })
-          .then((response) => {
-            return resolve(response.data);
-          })
-          .catch((error) => {
-            return reject(error);
-          });
-      });
-    } else {
-      // do nothing
-    }
-  }
-</script>
